Adds versatility sorting for Legion Remix
Allows users to sort the top 10 player list by either total Infinite Power or estimated Limits Unbound (versatility) when playing the Legion Remix event.

This provides an alternative sorting option that may be more relevant for some players during the event. It introduces a new setting in the options panel for selecting the sort criteria.